refactor(migrate): uses after_migrate hook for UUID setup
